+++ Source/Tam/Maestro/Services/Broadcast/Services.Broadcast/Entities/ProposalInventoryWeekDaypartDetailDto.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public string InventorySpot { get; set; }
         public string Contract { get; set; }
-        public float Eff { get; set; }
-        public float EffVariation { get; set; }
+        public double Eff { get; set; }
+        public double EffVariation { get; set; }
         public double Impressions { get; set; }
         public decimal Cpm { get; set; }
         public decimal Cost { get; set; }
