+++ Source/Tam/MaestroTests/SupportTests/Services.Broadcast.IntegrationTests/ApplicationServices/SchedulesReportServiceIntegrationTest.cs
@@ -141,7 +141,7 @@ namespace Services.Broadcast.IntegrationTests.ApplicationServices
             public string Affiliate { get; set; }
             public string ProgramName { get; set; }
             public DisplayDaypart DisplayDaypart { get; set; }
-            public double? Cost { get; set; }
+            public decimal? Cost { get; set; }
             public int? OrderedSpots { get; set; }
             public int? DeliveredSpots { get; set; }
             public double OrderedImpressions { get; set; }
